Simplify config setting and update nsdk and uisdk versions
