+++ network-sdk-sample/app/build.gradle
@@ -44,7 +44,7 @@ android {
 }
 
 dependencies {
-    api 'com.karhoo:nsdk:1.1.0'
+    api 'com.karhoo:nsdk:1.1.1'
     api 'com.braintreepayments.api:drop-in:4.4.0'
 
     implementation fileTree(dir: "libs", include: ["*.jar"])
