+++ network-sdk-sample/app/src/main/java/com/karhoo/samples/networksdk/config/KarhooSandboxConfig.kt
@@ -1,0 +1,29 @@
+package com.karhoo.samples.networksdk.config
+
+import android.content.Context
+import com.karhoo.sdk.analytics.AnalyticProvider
+import com.karhoo.sdk.api.KarhooEnvironment
+import com.karhoo.sdk.api.KarhooSDKConfiguration
+import com.karhoo.sdk.api.model.AuthenticationMethod
+
+class KarhooSandboxConfig(
+    private val context: Context,
+    private val authMethod: AuthenticationMethod
+) : KarhooSDKConfiguration {
+
+    override fun context(): Context {
+        return context
+    }
+
+    override fun environment(): KarhooEnvironment {
+        return KarhooEnvironment.Sandbox()
+    }
+
+    override fun analyticsProvider(): AnalyticProvider? {
+        return null
+    }
+
+    override fun authenticationMethod(): AuthenticationMethod {
+        return authMethod
+    }
+}
